refactor(builder): extract measurement and api modules, simplify wheel handler
- Extract Mode B measurement subsystem to measurement.ts (ResizeObserver,
  scroll correction, content size deferral, stayAtEnd)
- Extract public API assembly to api.ts (data wrappers, scroll methods,
  event subscription, destroy teardown)
- Simplify wheel handler: delegate to onScrollFrame instead of
  reimplementing scroll logic (fixes missing velocity:change and
  afterScroll events during wheel scroll)
- Add re-entry guard in onScrollFrame to skip redundant native scroll
  events after programmatic scrollTop assignment
- Deduplicate idle timeout handler (onScrollIdle + scheduleIdle)
- Deduplicate click/dblclick traversal (findClickTarget)
- Use positional params + short property keys for minification

core.ts: 1513 → 1097 lines (-28%)
Base minified: 24.1 → 23.6 KB (-0.5 KB)
Base gzipped: 8.7 → 8.8 KB (+0.1 KB)
Tests: 2763/2763 pass, typecheck clean
